Fix controller declarations and operators
in controller.cpp there is a problem using the << operator in log_debug. generating the error: no type named 'type' in 'struct std::enable_if<false, void>', the operator<< is causing a substitution failure.
in controller.hpp there is also a problem in the declaration of the get_ff_features() function in the Controller class in the controller.hpp file, generating the error: no match for 'operator<<' (operand types are 'std::basic_ostream<char>' and 'boost::function<void()>').
there is also the error: 'vector' in namespace 'std' does not name a template type, where the compiler does not recognize the std::vector type because you did not include the <vector> header in the file where std::vector is used.

problems were found using gcc-c++ version 13